Name synth handle helpers

+++ src/main/audio/synth_internal.h
@@ -340,7 +340,7 @@ void synthSetFadeAction(u32 fadeIndex, u8 action);
 void synthQueueVoice(SynthVoice* voice);
 void synthQueueHandle(u32 handle);
 void synthFreeHandle(u32 handle);
-void synthSetHandleValue16(u32 handle, u16 value);
+void synthSetHandleValue16(u32 handle, u32 value);
 void synthRestoreQueuedHandle(u32 handle);
 void synthSetHandleMixData(u32 handle, u32 value0, u32 value1);
 void synthSetControllerValue(u8 controller, u8 studioIndex, u8 channelIndex, u8 value);
